fix: js package
