+++ packages/config-eslint/library.js
@@ -30,5 +30,5 @@ module.exports = {
       },
     },
   },
-  ignorePatterns: ["node_modules/", "dist/", "*.config.js"],
+  ignorePatterns: ["node_modules/", "dist/", "*.config.js", "*.config.ts"],
 };
